+++ langchain/src/hub/tests/hub.int.test.ts
@@ -112,3 +112,14 @@ test("Test LangChain Hub while loading model with dynamic imports and structured
   expect(res).not.toBeInstanceOf(AIMessage);
   expect(typeof res.correctness).toBe("boolean");
 });
+
+test("Test LangChain Hub while loading model not defined in a RunnableBinding", async () => {
+  const promptA = await nodePull("hntrl/binding-manifest", {
+    includeModel: true,
+  });
+  const resA = await promptA.invoke({
+    question: "What's the capital of the USA?",
+  });
+  expect(resA).toBeInstanceOf(AIMessage);
+  expect(resA.content).toBeDefined();
+});