+++ langchain-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @langchain/core
 
+## 0.3.75
+
+### Patch Changes
+
+- d6d841f: fix(core): Fix deep nesting of runnables within traceables
+
+## 0.3.74
+
+### Patch Changes
+
+- 4e53005: fix(core): Always inherit parent run id onto callback manager from context
+
 ## 0.3.73
 
 ### Patch Changes